Move test fixtures into conftest.py
Why these changes are being introduced:
* If a new test file is added this allows the test fixtures to be shared across all tests in the directory.
How this addresses that need:
* moves test fixture into conftest.py
